Create executor/read and move stuff in
