+++ executor/tests/execute_expression.rs
@@ -13,7 +13,7 @@ use compiler::annotation::expression::{
     ExpressionCompileError,
 };
 use encoding::value::{value::Value, value_type::ValueTypeCategory};
-use executor::expression_executor::{ExpressionExecutor, ExpressionValue};
+use executor::read::expression_executor::{ExpressionExecutor, ExpressionValue};
 use ir::{
     pattern::constraint::Constraint,
     pipeline::{function_signature::HashMapFunctionSignatureIndex, ParameterRegistry},
